Generalized, but this approach looks infeasible.
Tried to generalize with arguments, but it looks like this will take a
very long time without a smarter plan.

We need to find some optimizations to make, such as seeing if there's
some repeated essential configuration, such as a combination of the
contours of the stage, the current piece, and the next move.