Got jobs to parse, but unsatisfied by how whitespace is handled.
I'm not quite enamored with nom with this experience.  One major
problem is mixing concerns of tokens versus structure; it feels like
whitespace, which is a token concern, infects all of the parser
functions, and it's not that fun.  Will need to think a bit more
before I'm completely against this.